+++ src/test/java/br/com/estoque/controller/InsumoControllerTest.java
@@ -16,21 +16,29 @@ import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-
-
+import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
+import br.com.estoque.security.SecurityFilter;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
 
 import br.com.estoque.entity.Insumo;
 import br.com.estoque.exception.GlobalHandlerException;
 import br.com.estoque.exception.ResourceNotFoundException;
 import br.com.estoque.service.InsumoService;
 
-@WebMvcTest(InsumoController.class)
+@WebMvcTest(
+		 controllers = InsumoController.class,
+	        excludeFilters = @ComponentScan.Filter(
+	                type = FilterType.ASSIGNABLE_TYPE,
+	                classes = SecurityFilter.class
+	        ))
 @Import(GlobalHandlerException.class)
+@AutoConfigureMockMvc(addFilters = false)
 public class InsumoControllerTest {
 	
 	@Autowired